% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clean.R
\name{is_valid_word}
\alias{is_valid_word}
\title{¿La palabra contiene solo letras válidas?}
\usage{
is_valid_word(word, lang = syll_get_option("lang"))
}
\arguments{
\item{word}{Cadena de texto (longitud 1).}

\item{lang}{Código de idioma: \code{"es"} o \code{"ca"}.}
}
\value{
\code{TRUE}/\code{FALSE} o \code{NA} si \code{word} es inválida.
}
\description{
Verifica que todos los caracteres pertenezcan al alfabeto interno soportado
para el idioma indicado.
}
\examples{
is_valid_word("pingüino")             # TRUE
is_valid_word("pa\u00efsos", lang = "ca")  # TRUE
is_valid_word("hola!")                # FALSE
}
\seealso{
Other syllabify:
\code{\link[=clean_word]{clean_word()}},
\code{\link[=syll_count]{syll_count()}},
\code{\link[=syll_hyphenate]{syll_hyphenate()}},
\code{\link[=syll_split]{syll_split()}},
\code{\link[=tokenize_clean]{tokenize_clean()}}
}
\concept{syllabify}
