% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syll-core.R
\name{syll_hyphenate}
\alias{syll_hyphenate}
\title{Inserta separadores entre sílabas}
\usage{
syll_hyphenate(word, lang = syll_get_option("lang"), hyphen = "-")
}
\arguments{
\item{word}{Cadena (longitud 1).}

\item{lang}{Código de idioma: \code{"es"} o \code{"ca"}. Por defecto,
\code{syll_get_option("lang")}.}

\item{hyphen}{Cadena usada como separador (por defecto \code{"-"}).}
}
\value{
Cadena con separadores entre sílabas, o \code{NA} si no se pudo
segmentar.
}
\description{
Útil para visualización o verificación de silabificación.
}
\examples{
syll_hyphenate("silabificación") # "si-la-bi-fi-ca-ción"
syll_hyphenate("agradar-me", lang = "ca")  # "a-gra-dar-me"
}
\seealso{
Other syllabify:
\code{\link[=clean_word]{clean_word()}},
\code{\link[=is_valid_word]{is_valid_word()}},
\code{\link[=syll_count]{syll_count()}},
\code{\link[=syll_split]{syll_split()}},
\code{\link[=tokenize_clean]{tokenize_clean()}}
}
\concept{syllabify}
