% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{syllabify-internals}
\alias{syllabify-internals}
\alias{.valid_letters}
\alias{.v_fuerte}
\alias{.v_debil}
\alias{.v_tilde}
\alias{.clusters_lr}
\alias{.abbr_es}
\alias{.valid_letters_ca}
\alias{.v_fuerte_ca}
\alias{.v_debil_ca}
\alias{.v_tilde_ca}
\alias{.digraphs_ca}
\alias{.clusters_ca}
\alias{.rx_valid_letters}
\alias{.rx_vowel}
\alias{.rx_strong}
\alias{.rx_weak}
\alias{.syll_cache}
\alias{.abbr_ca}
\title{Internos para segmentación silábica}
\usage{
.valid_letters

.v_fuerte

.v_debil

.v_tilde

.clusters_lr

.abbr_es

.valid_letters_ca

.v_fuerte_ca

.v_debil_ca

.v_tilde_ca

.digraphs_ca

.clusters_ca

.rx_valid_letters

.rx_vowel

.rx_strong

.rx_weak

.syll_cache

.abbr_ca
}
\description{
Conjunto de reglas, constantes y expresiones regulares precompiladas
empleadas por el motor de segmentación silábica en español y catalán.
}
\details{
Estos objetos se inicializan en \code{.onLoad()} y no deben modificarse
fuera de ese hook.
}
\section{Contenido (español)}{

\itemize{
\item \code{.valid_letters}: letras válidas en minúsculas (incluye tildes, ñ y diéresis).
\item \code{.v_fuerte}: vocales fuertes.
\item \code{.v_debil}: vocales débiles (incluye \code{ï}/\code{ü} para marcar no tónicas).
\item \code{.v_tilde}: vocales acentuadas.
\item \code{.clusters_lr}: grupos consonánticos que no se separan (combinaciones con l o r).
}
}

\section{Contenido (catalán)}{

\itemize{
\item \code{.valid_letters_ca}: letras válidas en minúsculas para catalán.
\item \code{.v_fuerte_ca}: vocales fuertes en catalán (a, e, o con y sin acento/grave).
\item \code{.v_debil_ca}: vocales débiles en catalán (i, u sin acento; ï, ü con diéresis).
\item \code{.v_tilde_ca}: vocales acentuadas en catalán (à, è, é, í, ï, ò, ó, ú, ü).
\item \code{.digraphs_ca}: dígrafos inseparables propios del catalán (ll, ny, rr, ss, gu, qu).
\item \code{.clusters_ca}: grupos consonánticos inseparables en catalán (bl, br, cl, cr, ...).
}
}

\keyword{internal}
