% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean.R
\name{is_valid_word}
\alias{is_valid_word}
\title{¿La palabra contiene solo letras válidas?}
\usage{
is_valid_word(word)
}
\arguments{
\item{word}{Cadena de texto (longitud 1).}
}
\value{
\code{TRUE}/\code{FALSE} o \code{NA} si \code{word} es inválida.
}
\description{
Verifica que todos los caracteres pertenezcan al alfabeto interno soportado.
}
\examples{
is_valid_word("pingüino")  # TRUE
is_valid_word("hola!")     # FALSE
}
\seealso{
Other syllabify-es: 
\code{\link{clean_word}()},
\code{\link{syll-cache}},
\code{\link{syll-config}},
\code{\link{syll_count}()},
\code{\link{syll_hyphenate}()},
\code{\link{syll_split}()},
\code{\link{syllabify-internals}},
\code{\link{tokenize_clean}()}
}
\concept{syllabify-es}
