% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{syll-config}
\alias{syll-config}
\alias{syll_config}
\alias{syll_get_option}
\alias{syll_set_options}
\alias{syll_reset_options}
\title{Configuración del paquete}
\usage{
syll_config()

syll_get_option(name, default = NULL)

syll_set_options(...)

syll_reset_options(which = "all")
}
\description{
Consulta y modificación de opciones de silabificación, incluyendo el
idioma global (\code{lang}).
}
\examples{
syll_get_option("lang")        # "es"
syll_set_options(lang = "ca")  # cambia a catalán globalmente
syll_reset_options()           # restaura todo a los valores por defecto

}
\keyword{internal}
