% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{syll-config}
\alias{syll-config}
\alias{syll_config}
\alias{syll_get_option}
\alias{syll_set_options}
\alias{syll_reset_options}
\title{Configuración del paquete (opciones de sílabas para ES)}
\usage{
syll_config()

syll_get_option(name, default = NULL)

syll_set_options(...)

syll_reset_options(which = "all")
}
\description{
Consulta y modificación de opciones de silabificación.
}
\seealso{
Other syllabify-es: 
\code{\link{clean_word}()},
\code{\link{is_valid_word}()},
\code{\link{syll-cache}},
\code{\link{syll_count}()},
\code{\link{syll_hyphenate}()},
\code{\link{syll_split}()},
\code{\link{syllabify-internals}},
\code{\link{tokenize_clean}()}
}
\concept{syllabify-es}
\keyword{internal}
