% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{syllabify-internals}
\alias{syllabify-internals}
\alias{.valid_letters}
\alias{.v_fuerte}
\alias{.v_debil}
\alias{.v_tilde}
\alias{.clusters_lr}
\alias{.rx_valid_letters}
\alias{.rx_vowel}
\alias{.rx_strong}
\alias{.rx_weak}
\alias{.syll_cache}
\title{Internos para segmentación silábica en español}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 6.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 12.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{environment} of length 0.
}
\usage{
.valid_letters

.v_fuerte

.v_debil

.v_tilde

.clusters_lr

.rx_valid_letters

.rx_vowel

.rx_strong

.rx_weak

.syll_cache
}
\description{
Conjunto de reglas, constantes y expresiones regulares precompiladas
empleadas por el motor de segmentación silábica en español.
}
\details{
Estos objetos se inicializan en \code{.onLoad()} y no deben modificarse
fuera de ese hook.
}
\section{Contenido}{

\itemize{
\item \code{.valid_letters}: letras válidas en minúsculas (incluye tildes, ñ y diéresis).
\item \code{.v_fuerte}: vocales fuertes.
\item \code{.v_debil}: vocales débiles (incluye \code{ï}/\code{ü} para marcar no tónicas).
\item \code{.v_tilde}: vocales acentuadas.
\item \code{.clusters_lr}: grupos consonánticos que no se separan (combinaciones con l o r, p. ej., \code{bl}, \code{br}, \code{cl}, \code{cr}, ...).
\item \code{.rx_valid_letters}, \code{.rx_vowel}, \code{.rx_strong}, \code{.rx_weak}: regex precompiladas.
\item \code{.syll_cache}: entorno para cachear segmentaciones.
}
}

\seealso{
\code{.onLoad()}

Other syllabify-es: 
\code{\link{clean_word}()},
\code{\link{is_valid_word}()},
\code{\link{syll-cache}},
\code{\link{syll-config}},
\code{\link{syll_count}()},
\code{\link{syll_hyphenate}()},
\code{\link{syll_split}()},
\code{\link{tokenize_clean}()}
}
\concept{syllabify-es}
\keyword{datasets}
\keyword{internal}
